refactor: make observability package public
This is to allow for Zigflow to be used programmatically
